+++ apps/ledger-live-desktop/src/renderer/components/TopBar/NotificationIndicator/InformationDrawer.tsx
@@ -11,7 +11,7 @@ import { useSelector, useDispatch } from "react-redux";
 import { informationCenterStateSelector } from "~/renderer/reducers/UI";
 import { setTabInformationCenter } from "~/renderer/actions/UI";
 import { useFilteredServiceStatus } from "@ledgerhq/live-common/notifications/ServiceStatusProvider/index";
-import { notificationsContentCardSelector } from "~/renderer/reducers/dynamicContent";
+import { useUnseenNotificationsCount } from "~/renderer/hooks/useUnseenNotificationsCount";
 
 const FADE_DURATION = 200;
 const PanelContainer = styled.div`
@@ -39,8 +39,7 @@ export const InformationDrawer = ({
 }) => {
   const { t } = useTranslation();
   const { incidents } = useFilteredServiceStatus();
-  const notificationsCards = useSelector(notificationsContentCardSelector);
-  const unseenCount = notificationsCards?.filter(n => !n.viewed).length || 0;
+  const unseenCount = useUnseenNotificationsCount();
   const incidentCount = incidents.length;
   const { tabId } = useSelector(informationCenterStateSelector);