+++ apps/ledger-live-mobile/src/screens/Platform/exchange/StartExchange.tsx
@@ -44,7 +44,7 @@ export default function PlatformStartExchange({ navigation, route }: Props) {
   const requestToSetHeaderOptions = useCallback(() => undefined, []);
   const request = useMemo(() => route.params.request, [route.params.request]);
   return (
-    <SafeAreaView style={styles.root}>
+    <SafeAreaView style={styles.root} edges={["bottom"]}>
       <Flex px={16} py={8} flex={1}>
         <SelectDevice2
           onSelect={setDevice}
@@ -63,9 +63,4 @@ export default function PlatformStartExchange({ navigation, route }: Props) {
   );
 }
 
-const styles = StyleSheet.create({
-  root: {
-    flex: 1,
-    padding: 32,
-  },
-});
+const styles = StyleSheet.create({ root: { flex: 1 } });