+++ libs/ledgerjs/packages/types-live/src/lnsUpsell.ts
@@ -2,8 +2,7 @@ export type LldNanoSUpsellBannersConfig = {
   manager: boolean;
   accounts: boolean;
   notification_center: boolean;
-  img: string;
   link: string;
-  learn_more: string;
-  "%": number;
+  img?: string;
+  "%"?: number;
 };