+++ libs/coin-modules/coin-evm/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ledgerhq/coin-evm
 
+## 2.14.0-next.0
+
+### Minor Changes
+
+- [#9446](https://github.com/LedgerHQ/ledger-live/pull/9446) [`dcea0f8`](https://github.com/LedgerHQ/ledger-live/commit/dcea0f880728c808fffe5b09410eec614083f04b) Thanks [@Justkant](https://github.com/Justkant)! - fix: potential wrong type for the hydrate value
+
+- [#9472](https://github.com/LedgerHQ/ledger-live/pull/9472) [`9c4c82e`](https://github.com/LedgerHQ/ledger-live/commit/9c4c82e8e81aaadfc1132f0a25812be094e54453) Thanks [@KVNLS](https://github.com/KVNLS)! - Fix elliptic library version used
+
+### Patch Changes
+
+- Updated dependencies [[`9c4c82e`](https://github.com/LedgerHQ/ledger-live/commit/9c4c82e8e81aaadfc1132f0a25812be094e54453), [`cc00249`](https://github.com/LedgerHQ/ledger-live/commit/cc002495f3e107aba283a3aa4abca90954de6d76)]:
+  - @ledgerhq/evm-tools@1.6.0-next.0
+  - @ledgerhq/coin-framework@2.3.0-next.0
+  - @ledgerhq/domain-service@1.2.22-next.0
+
 ## 2.13.0
 
 ### Minor Changes